feat: deleted central dtr search and outdated models and methods related
