+++ consumer-backend/productpass/src/main/java/org/eclipse/tractusx/productpass/models/dtregistry/DigitalTwin.java
@@ -23,6 +23,7 @@
 
 package org.eclipse.tractusx.productpass.models.dtregistry;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -30,19 +31,26 @@ import com.fasterxml.jackson.databind.JsonNode;
 import java.util.ArrayList;
 
 /**
- * This class consists exclusively to define attributes related to the designed model of the Digital Twin's first version.
+ * This class consists exclusively to define attributes related to the designed model of the Digital Twin's improved version.
  **/
+@JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class DigitalTwin {
 
     /** ATTRIBUTES **/
     @JsonProperty("description")
     ArrayList<JsonNode> description;
-    @JsonProperty("globalAssetId")
-    JsonNode globalAssetId;
     @JsonProperty("idShort")
     String idShort;
-    @JsonProperty("identification")
+    @JsonProperty("assetKind")
+    String assetKind;
+    @JsonProperty("assetType")
+    String assetType;
+    @JsonProperty("globalAssetId")
+    String globalAssetId;
+    @JsonProperty("displayName")
+    Object displayName;
+    @JsonProperty("id")
     String identification;
     @JsonProperty("specificAssetIds")
     ArrayList<JsonNode> specificAssetIds;
@@ -54,10 +62,40 @@ public class DigitalTwin {
     public DigitalTwin() {
     }
     @SuppressWarnings("Unused")
-    public DigitalTwin(ArrayList<JsonNode> description, JsonNode globalAssetId, String idShort, String identification, ArrayList<JsonNode> specificAssetIds, ArrayList<SubModel> submodelDescriptors) {
+    public DigitalTwin(ArrayList<JsonNode> description, String idShort, String identification, ArrayList<JsonNode> specificAssetIds, ArrayList<SubModel> submodelDescriptors) {
         this.description = description;
-        this.globalAssetId = globalAssetId;
         this.idShort = idShort;
+        this.identification = identification;
+        this.specificAssetIds = specificAssetIds;
+        this.submodelDescriptors = submodelDescriptors;
+    }
+    @SuppressWarnings("Unused")
+    public DigitalTwin(ArrayList<JsonNode> description, String idShort, Object displayName, String identification, ArrayList<JsonNode> specificAssetIds, ArrayList<SubModel> submodelDescriptors) {
+        this.description = description;
+        this.idShort = idShort;
+        this.displayName = displayName;
+        this.identification = identification;
+        this.specificAssetIds = specificAssetIds;
+        this.submodelDescriptors = submodelDescriptors;
+    }
+    @SuppressWarnings("Unused")
+    public DigitalTwin(ArrayList<JsonNode> description, String idShort, String globalAssetId, Object displayName, String identification, ArrayList<JsonNode> specificAssetIds, ArrayList<SubModel> submodelDescriptors) {
+        this.description = description;
+        this.idShort = idShort;
+        this.globalAssetId = globalAssetId;
+        this.displayName = displayName;
+        this.identification = identification;
+        this.specificAssetIds = specificAssetIds;
+        this.submodelDescriptors = submodelDescriptors;
+    }
+    @SuppressWarnings("Unused")
+    public DigitalTwin(ArrayList<JsonNode> description, String idShort, String assetKind, String assetType, String globalAssetId, Object displayName, String identification, ArrayList<JsonNode> specificAssetIds, ArrayList<SubModel> submodelDescriptors) {
+        this.description = description;
+        this.idShort = idShort;
+        this.assetKind = assetKind;
+        this.assetType = assetType;
+        this.globalAssetId = globalAssetId;
+        this.displayName = displayName;
         this.identification = identification;
         this.specificAssetIds = specificAssetIds;
         this.submodelDescriptors = submodelDescriptors;
@@ -69,14 +107,6 @@ public class DigitalTwin {
     }
     public void setDescription(ArrayList<JsonNode> description) {
         this.description = description;
-    }
-    @SuppressWarnings("Unused")
-    public JsonNode getGlobalAssetId() {
-        return globalAssetId;
-    }
-    @SuppressWarnings("Unused")
-    public void setGlobalAssetId(JsonNode globalAssetId) {
-        this.globalAssetId = globalAssetId;
     }
     public String getIdShort() {
         return idShort;
@@ -106,5 +136,36 @@ public class DigitalTwin {
     public void setSubmodelDescriptors(ArrayList<SubModel> submodelDescriptors) {
         this.submodelDescriptors = submodelDescriptors;
     }
+    @SuppressWarnings("Unused")
+    public Object getDisplayName() {
+        return displayName;
+    }
+    @SuppressWarnings("Unused")
+    public void setDisplayName(Object displayName) {
+        this.displayName = displayName;
+    }
+    @SuppressWarnings("Unused")
+    public String getGlobalAssetId() {
+        return globalAssetId;
+    }
+    @SuppressWarnings("Unused")
+    public void setGlobalAssetId(String globalAssetId) {
+        this.globalAssetId = globalAssetId;
+    }
+    @SuppressWarnings("Unused")
+    public String getAssetKind() {
+        return assetKind;
+    }
+    @SuppressWarnings("Unused")
+    public void setAssetKind(String assetKind) {
+        this.assetKind = assetKind;
+    }
+    @SuppressWarnings("Unused")
+    public String getAssetType() {
+        return assetType;
+    }
+    @SuppressWarnings("Unused")
+    public void setAssetType(String assetType) {
+        this.assetType = assetType;
+    }
 }
-
