+++ consumer-backend/productpass/src/main/java/org/eclipse/tractusx/productpass/models/dtregistry/EndPoint.java
@@ -23,19 +23,27 @@
 
 package org.eclipse.tractusx.productpass.models.dtregistry;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.google.common.base.Splitter;
+import utils.LogUtil;
+
+import java.util.List;
+import java.util.Map;
 
 /**
- * This class consists exclusively to define attributes related to the designed model of the Endpoint's first version.
+ * This class consists exclusively to define attributes related to the designed model of the Endpoint's improved version.
  **/
+@JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class EndPoint {
 
     /** ATTRIBUTES **/
     @JsonProperty("interface")
     String interfaceName;
-    ProtocolInformation protocolInformation = new ProtocolInformation();
+    @JsonProperty("protocolInformation")
+    ProtocolInformation3 protocolInformation = new ProtocolInformation3();
 
     /** GETTERS AND SETTERS **/
     public String getInterfaceName() {
@@ -45,38 +53,74 @@ public class EndPoint {
     public void setInterfaceName(String interfaceName) {
         this.interfaceName = interfaceName;
     }
-    public ProtocolInformation getProtocolInformation() {
+    public ProtocolInformation3 getProtocolInformation() {
         return protocolInformation;
     }
     @SuppressWarnings("Unused")
-    public void setProtocolInformation(String endpointAddress, String endpointProtocol, String endpointProtocolVersion) {
-        this.protocolInformation = new ProtocolInformation(endpointAddress, endpointProtocol, endpointProtocolVersion);
+    public void setProtocolInformation(String endpointAddress, String endpointProtocol, List<String> endpointProtocolVersion) {
+        this.protocolInformation = new ProtocolInformation3(endpointAddress, endpointProtocol, endpointProtocolVersion);
     }
 
     /** INNER CLASSES **/
     /**
-     * This class consists exclusively to define attributes related to the needed protocol information for the first version.
+     * This class consists exclusively to define attributes related to the needed protocol information for the improved version.
      **/
-    public static class ProtocolInformation {
+    public static class ProtocolInformation3 {
 
         /** ATTRIBUTES **/
-        @JsonProperty("endpointAddress")
+        @JsonProperty("href")
         String endpointAddress;
         @JsonProperty("endpointProtocol")
         String endpointProtocol;
+        @JsonProperty("subprotocol")
+        String subprotocol;
+        @JsonProperty("subprotocolBody")
+        String subprotocolBody;
+        @JsonProperty("subprotocolBodyEncoding")
+        String subprotocolBodyEncoding;
+        @JsonProperty("securityAttributes")
+        Object securityAttributes;
         @JsonProperty("endpointProtocolVersion")
-        String endpointProtocolVersion;
+        List<String> endpointProtocolVersion;
 
         /** CONSTRUCTOR(S) **/
-        public ProtocolInformation(String endpointAddress, String endpointProtocol, String endpointProtocolVersion) {
+        public ProtocolInformation3(String endpointAddress, String endpointProtocol, List<String> endpointProtocolVersion) {
             this.endpointAddress = endpointAddress;
             this.endpointProtocol = endpointProtocol;
             this.endpointProtocolVersion = endpointProtocolVersion;
         }
-        public ProtocolInformation() {
+        public ProtocolInformation3() {
+        }
+        @SuppressWarnings("Unused")
+        public ProtocolInformation3(String endpointAddress, String endpointProtocol, String subprotocol, String subprotocolBody, List<String> endpointProtocolVersion) {
+            this.endpointAddress = endpointAddress;
+            this.endpointProtocol = endpointProtocol;
+            this.subprotocol = subprotocol;
+            this.subprotocolBody = subprotocolBody;
+            this.endpointProtocolVersion = endpointProtocolVersion;
+        }
+        @SuppressWarnings("Unused")
+        public ProtocolInformation3(String endpointAddress, String endpointProtocol, String subprotocol, String subprotocolBody, String subprotocolBodyEncoding, List<String> endpointProtocolVersion) {
+            this.endpointAddress = endpointAddress;
+            this.endpointProtocol = endpointProtocol;
+            this.subprotocol = subprotocol;
+            this.subprotocolBody = subprotocolBody;
+            this.subprotocolBodyEncoding = subprotocolBodyEncoding;
+            this.endpointProtocolVersion = endpointProtocolVersion;
+        }
+        @SuppressWarnings("Unused")
+        public ProtocolInformation3(String endpointAddress, String endpointProtocol, String subprotocol, String subprotocolBody, String subprotocolBodyEncoding, Object securityAttributes, List<String> endpointProtocolVersion) {
+            this.endpointAddress = endpointAddress;
+            this.endpointProtocol = endpointProtocol;
+            this.subprotocol = subprotocol;
+            this.subprotocolBody = subprotocolBody;
+            this.subprotocolBodyEncoding = subprotocolBodyEncoding;
+            this.securityAttributes = securityAttributes;
+            this.endpointProtocolVersion = endpointProtocolVersion;
         }
 
         /** GETTERS AND SETTERS **/
+        @SuppressWarnings("Unused")
         public String getEndpointAddress() {
             return endpointAddress;
         }
@@ -93,12 +137,52 @@ public class EndPoint {
             this.endpointProtocol = endpointProtocol;
         }
         @SuppressWarnings("Unused")
-        public String getEndpointProtocolVersion() {
+        public List<String> getEndpointProtocolVersion() {
             return endpointProtocolVersion;
         }
         @SuppressWarnings("Unused")
-        public void setEndpointProtocolVersion(String endpointProtocolVersion) {
+        public void setEndpointProtocolVersion(List<String> endpointProtocolVersion) {
             this.endpointProtocolVersion = endpointProtocolVersion;
+        }
+        @SuppressWarnings("Unused")
+        public String getSubprotocol() {
+            return subprotocol;
+        }
+        @SuppressWarnings("Unused")
+        public void setSubprotocol(String subprotocol) {
+            this.subprotocol = subprotocol;
+        }
+        @SuppressWarnings("Unused")
+        public String getSubprotocolBody() {
+            return subprotocolBody;
+        }
+        public Map<String, String> getParsedSubprotocolBody() {
+            try {
+                return Splitter.on(';').withKeyValueSeparator('=').split(this.subprotocolBody);
+            }catch (Exception e){
+                LogUtil.printException(e, "Error when parsing the subprotocol body params!");
+                return null;
+            }
+        }
+        @SuppressWarnings("Unused")
+        public void setSubprotocolBody(String subprotocolBody) {
+            this.subprotocolBody = subprotocolBody;
+        }
+        @SuppressWarnings("Unused")
+        public String getSubprotocolBodyEncoding() {
+            return subprotocolBodyEncoding;
+        }
+        @SuppressWarnings("Unused")
+        public void setSubprotocolBodyEncoding(String subprotocolBodyEncoding) {
+            this.subprotocolBodyEncoding = subprotocolBodyEncoding;
+        }
+        @SuppressWarnings("Unused")
+        public Object getSecurityAttributes() {
+            return securityAttributes;
+        }
+        @SuppressWarnings("Unused")
+        public void setSecurityAttributes(Object securityAttributes) {
+            this.securityAttributes = securityAttributes;
         }
     }
 }
