+++ consumer-backend/productpass/src/main/java/org/eclipse/tractusx/productpass/models/dtregistry/SubModel.java
@@ -23,14 +23,18 @@
 
 package org.eclipse.tractusx.productpass.models.dtregistry;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import java.util.ArrayList;
 
 /**
- * This class consists exclusively to define attributes related to the designed model of the Submodel's first version.
+ * This class consists exclusively to define attributes related to the designed model of the Submodel's improved version.
  **/
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(JsonInclude.Include.NON_NULL)
 public class SubModel {
 
     /** ATTRIBUTES **/
@@ -38,16 +42,21 @@ public class SubModel {
     ArrayList<JsonNode> description;
     @JsonProperty("idShort")
     String idShort;
-    @JsonProperty("identification")
+
+    @JsonProperty("supplementalSemanticId")
+    Object supplementalSemanticId;
+
+    @JsonProperty("id")
     String identification;
     @JsonProperty("semanticId")
-    JsonNode semanticId;
+    SemanticId semanticId;
+
     @JsonProperty("endpoints")
     ArrayList<EndPoint> endpoints;
 
     /** CONSTRUCTOR(S) **/
     @SuppressWarnings("Unused")
-    public SubModel(ArrayList<JsonNode> description, String idShort, String identification, JsonNode semanticId, ArrayList<EndPoint> endpoints) {
+    public SubModel(ArrayList<JsonNode> description, String idShort, String identification, SemanticId semanticId, ArrayList<EndPoint> endpoints) {
         this.description = description;
         this.idShort = idShort;
         this.identification = identification;
@@ -57,20 +66,33 @@ public class SubModel {
     @SuppressWarnings("Unused")
     public SubModel() {
     }
+    @SuppressWarnings("Unused")
+    public SubModel(ArrayList<JsonNode> description, String idShort, Object supplementalSemanticId, String identification, SemanticId semanticId, ArrayList<EndPoint> endpoints) {
+        this.description = description;
+        this.idShort = idShort;
+        this.supplementalSemanticId = supplementalSemanticId;
+        this.identification = identification;
+        this.semanticId = semanticId;
+        this.endpoints = endpoints;
+    }
 
     /** GETTERS AND SETTERS **/
     public ArrayList<JsonNode> getDescription() {
         return description;
     }
+
     public void setDescription(ArrayList<JsonNode> description) {
         this.description = description;
     }
+
     public String getIdShort() {
         return idShort;
     }
+
     public void setIdShort(String idShort) {
         this.idShort = idShort;
     }
+
     public String getIdentification() {
         return identification;
     }
@@ -78,16 +100,97 @@ public class SubModel {
     public void setIdentification(String identification) {
         this.identification = identification;
     }
-    public JsonNode getSemanticId() {
+
+    public SemanticId getSemanticId() {
         return semanticId;
     }
-    public void setSemanticId(JsonNode semanticId) {
+
+    public void setSemanticId(SemanticId semanticId) {
         this.semanticId = semanticId;
     }
+
     public ArrayList<EndPoint> getEndpoints() {
         return endpoints;
     }
+
     public void setEndpoints(ArrayList<EndPoint> endpoints) {
         this.endpoints = endpoints;
+    }
+    @SuppressWarnings("Unused")
+    public Object getSupplementalSemanticId() {
+        return supplementalSemanticId;
+    }
+    @SuppressWarnings("Unused")
+    public void setSupplementalSemanticId(Object supplementalSemanticId) {
+        this.supplementalSemanticId = supplementalSemanticId;
+    }
+
+    /** INNER CLASSES **/
+    /**
+     * This class consists exclusively to define attributes related to the Submodel's SemanticId.
+     **/
+    public static class SemanticId {
+
+        /** ATTRIBUTES **/
+        @JsonProperty("type")
+        String type;
+        @JsonProperty("keys")
+        ArrayList<Key> keys;
+
+        /** CONSTRUCTOR(S) **/
+        @SuppressWarnings("Unused")
+        public SemanticId(String type, ArrayList<Key> keys) {
+            this.type = type;
+            this.keys = keys;
+        }
+        @SuppressWarnings("Unused")
+        public SemanticId() {
+        }
+
+        /** GETTERS AND SETTERS **/
+        public String getType() {
+            return type;
+        }
+
+        public void setType(String type) {
+            this.type = type;
+        }
+
+        public ArrayList<Key> getKeys() {
+            return keys;
+        }
+        @SuppressWarnings("Unused")
+        public void setKeys(ArrayList<Key> keys) {
+            this.keys = keys;
+        }
+
+        /** INNER CLASSES **/
+        /**
+         * This class consists exclusively to define attributes related to the SemanticId's keys property.
+         **/
+        @JsonIgnoreProperties(ignoreUnknown = true)
+        @JsonInclude(JsonInclude.Include.NON_NULL)
+        static public class Key {
+
+            /** ATTRIBUTES **/
+            @JsonProperty("type")
+            String type;
+            @JsonProperty("value")
+            String value;
+
+            /** CONSTRUCTOR(S) **/
+            public Key() {}
+
+            public Key(String type, String value) {
+                this.type = type;
+                this.value = value;
+            }
+
+            /** GETTERS AND SETTERS **/
+            public String getType() { return type; }
+            public void setType(String type) { this.type = type; }
+            public String getValue() { return value; }
+            public void setValue(String value) { this.value = value; }
+        }
     }
 }
