+++ consumer-backend/productpass/src/main/java/org/eclipse/tractusx/productpass/services/AasService.java
@@ -25,14 +25,14 @@ package org.eclipse.tractusx.productpass.services;
 
 import org.eclipse.tractusx.productpass.config.DtrConfig;
 import org.eclipse.tractusx.productpass.config.PassportConfig;
-import org.eclipse.tractusx.productpass.exceptions.ControllerException;
 import org.eclipse.tractusx.productpass.exceptions.ServiceException;
 import org.eclipse.tractusx.productpass.exceptions.ServiceInitializationException;
 import org.eclipse.tractusx.productpass.managers.DtrSearchManager;
 import org.eclipse.tractusx.productpass.managers.ProcessManager;
 import org.eclipse.tractusx.productpass.models.auth.JwtToken;
 import org.eclipse.tractusx.productpass.models.catenax.Dtr;
-import org.eclipse.tractusx.productpass.models.dtregistry.*;
+import org.eclipse.tractusx.productpass.models.dtregistry.DigitalTwin;
+import org.eclipse.tractusx.productpass.models.dtregistry.SubModel;
 import org.eclipse.tractusx.productpass.models.edc.AssetSearch;
 import org.eclipse.tractusx.productpass.models.edc.DataPlaneEndpoint;
 import org.eclipse.tractusx.productpass.models.http.requests.Search;
@@ -60,11 +60,8 @@ import java.util.Map;
 @Service
 public class AasService extends BaseService {
 
-    private final String submodelTypeKey = "Submodel";
     /** ATTRIBUTES **/
-    public String registryUrl;
-
-    public Boolean central;
+    private final String submodelTypeKey = "Submodel";
     private final HttpUtil httpUtil;
     private final JsonUtil jsonUtil;
     private final DtrConfig dtrConfig;
@@ -97,8 +94,6 @@ public class AasService extends BaseService {
      * Initiates the main needed variables for the AasService by loading from the application's configuration file.
      **/
     public void init(Environment env) {
-        this.registryUrl = dtrConfig.getCentralUrl();
-        this.central = dtrConfig.getCentral();
         Object decentralApis = dtrConfig.getDecentralApis();
         if (decentralApis == null) { // If the configuration is null use the default variables
             decentralApis = Map.of(
@@ -108,11 +103,6 @@ public class AasService extends BaseService {
             );
         }
         this.apis = Map.of(
-                "central", Map.of(
-                        "search", "/lookup/shells",
-                        "digitalTwin", "/registry/shell-descriptors",
-                        "subModel", "/submodel-descriptors"
-                ),
                 "decentral", decentralApis
         );
     }
@@ -126,92 +116,7 @@ public class AasService extends BaseService {
      */
     @Override
     public List<String> getEmptyVariables() {
-        List<String> missingVariables = new ArrayList<>();
-
-        if (this.central == null) {
-            missingVariables.add("dtr.central");
-        } else {
-            if (this.central) {
-                if (this.registryUrl.isEmpty()) {
-                    missingVariables.add("dtr.centralUrl");
-                }
-            }
-        }
-        return missingVariables;
-    }
-
-    /**
-     * Searches a {@code Submodel} from a Digital Twin by a given position index.
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin} object with its submodels.
-     * @param   position
-     *          the {@code Integer} position index of the intended submodel.
-     *
-     * @return a {@code Submodel} object from the position index, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a submodel for the specified position index.
-     */
-    @SuppressWarnings("Unused")
-    public SubModel searchSubModelInDigitalTwinByIndex(DigitalTwin digitalTwin, Integer position) {
-        try {
-            SubModel subModel = this.getSubModelFromDigitalTwin(digitalTwin, position);
-            if (subModel == null) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchSubModelInDigitalTwin",
-                        "It was not possible to get digital twin in the selected position for the selected asset type and the the selected assetId");
-            }
-            return subModel;
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getIdShort",
-                    e,
-                    "It was not possible to get subModel!");
-        }
-    }
-
-    /**
-     * Searches a {@code DigitalTwin3} from a given position index containing the specified parameters.
-     * <p>
-     * @param   assetType
-     *          the {@code String} type of the asset.
-     * @param   assetId
-     *          the {@code String} identification of the asset.
-     * @param   position
-     *          the {@code Integer} position index of the intended digital twin.
-     * @param   registryUrl
-     *          the {@code String} registry URL of the Digital Twin.
-     * @param   edr
-     *          the {@code DataPlaneEndpoint} object with the data plane needed parameters.
-     *
-     * @return a {@code DigitalTwin3} object from the position index, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a {@code DigitalTwin3} for the specified position index.
-     */
-    public DigitalTwin3 searchDigitalTwin3(String assetType, String assetId, Integer position, String registryUrl, DataPlaneEndpoint edr) {
-        try {
-            ArrayList<String> digitalTwinIds = this.queryDigitalTwin(assetType, assetId, registryUrl, edr);
-            if (digitalTwinIds == null || digitalTwinIds.size() == 0) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchDigitalTwin",
-                        "It was not possible to get digital twin for the selected asset type and the the selected assetId");
-            }
-            if (position > digitalTwinIds.size()) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchDigitalTwin",
-                        "It was not possible to get digital twin in the selected position for the selected asset type and the the selected assetId");
-            }
-
-            String digitalTwinId = digitalTwinIds.get(position);
-            DigitalTwin3 digitalTwin = this.getDigitalTwin3(digitalTwinId, registryUrl, edr);
-            if (digitalTwin == null) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchDigitalTwin",
-                        "It was not possible to get digital twin in the selected position for the selected asset type and the the selected assetId");
-            }
-            return digitalTwin;
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "searchDigitalTwin",
-                    e,
-                    "It was not possible to search digital twin!");
-        }
+        return new ArrayList<>(); //Without central search there's no missing variables
     }
 
     /**
@@ -245,7 +150,6 @@ public class AasService extends BaseService {
                         "It was not possible to get digital twin in the selected position for the selected asset type and the the selected assetId");
             }
 
-
             String digitalTwinId = digitalTwinIds.get(position);
             DigitalTwin digitalTwin = this.getDigitalTwin(digitalTwinId, registryUrl, edr);
             if (digitalTwin == null) {
@@ -272,112 +176,28 @@ public class AasService extends BaseService {
      *
      */
     public String getPathEndpoint(String registryUrl, String key) {
-        if (this.central || registryUrl == null || registryUrl.isEmpty()) {
-            return (String) jsonUtil.getValue(this.apis, "central." + key, ".", null);
-        }
         String path = (String) jsonUtil.getValue(this.apis, "decentral." + key, ".", null);
         return path;
     }
 
     /**
-     * Searches a {@code Submodel3} from a Digital Twin by a given short identification.
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin3} object with its submodels.
-     * @param   idShort
-     *          the {@code String} short id of the intended submodel.
-     *
-     * @return a {@code Submodel3} object with the submodel found, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a the {@code Submodel3} for the given id.
-     */
-    public SubModel3 searchSubModel3ById(DigitalTwin3 digitalTwin, String idShort) {
-        try {
-            SubModel3 subModel = this.getSubModel3ById(digitalTwin, idShort);
-            if (subModel == null) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchSubModel3ById",
-                        "It was not possible to get submodel in the selected position for the selected asset type and the the selected assetId");
-            }
-            return subModel;
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "searchSubModel3ById",
-                    e,
-                    "It was not possible to search submodel!");
-        }
-    }
-
-    /**
-     * Searches a {@code Submodel3} from a Digital Twin by a given semantic identification.
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin3} object with its submodels.
-     * @param   semanticId
-     *          the {@code String} semantic id of the intended submodel.
-     *
-     * @return a {@code Submodel3} object with the submodel found, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a the {@code Submodel3} for the given semantic id.
-     */
-    public SubModel3 searchSubModel3BySemanticId(DigitalTwin3 digitalTwin, String semanticId) {
-        try {
-            SubModel3 subModel = this.getSubModel3BySemanticId(digitalTwin, semanticId);
-            if (subModel == null) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchSubModel3BySemanticId",
-                        "It was not possible to get submodel in the selected position for the selected asset type and the the selected assetId");
-            }
-            return subModel;
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "searchSubModel3ById",
-                    e,
-                    "It was not possible to search submodel!");
-        }
-    }
-    /**
-     * Searches a {@code Submodel3} from a Digital Twin by the configured semanticIds
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin3} object with its submodels.
-     *
-     * @return a {@code Submodel3} object with the submodel found, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a the {@code Submodel3} for the given semantic id.
-     */
-    public SubModel3 searchSubModel3BySemanticId(DigitalTwin3 digitalTwin) {
-        try {
-            SubModel3 subModel = this.getSubModel3BySemanticId(digitalTwin);
-            if (subModel == null) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchSubModel3BySemanticId",
-                        "It was not possible to get submodel in the selected position for the selected asset type and the the selected assetId");
-            }
-            return subModel;
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "searchSubModel3ById",
-                    e,
-                    "It was not possible to search submodel!");
-        }
-    }
-
-    /**
-     * Searches a {@code Submodel} from a Digital Twin by a given short identification.
+     * Searches a {@code Submodel} from a Digital Twin by a given semantic identification.
      * <p>
      * @param   digitalTwin
      *          the {@code DigitalTwin} object with its submodels.
-     * @param   idShort
-     *          the {@code String} short id of the intended submodel.
+     * @param   semanticId
+     *          the {@code String} semantic id of the intended submodel.
      *
      * @return a {@code Submodel} object with the submodel found, if exists.
      *
      * @throws  ServiceException
-     *           if unable to find a the {@code Submodel} for the given id.
+     *           if unable to find a the {@code Submodel} for the given semantic id.
      */
-    public SubModel searchSubModelById(DigitalTwin digitalTwin, String idShort) {
+    public SubModel searchSubModelBySemanticId(DigitalTwin digitalTwin, String semanticId) {
         try {
-            SubModel subModel = this.getSubModelById(digitalTwin, idShort);
+            SubModel subModel = this.getSubModelBySemanticId(digitalTwin, semanticId);
             if (subModel == null) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchSubModelById",
+                throw new ServiceException(this.getClass().getName() + "." + "searchSubModelBySemanticId",
                         "It was not possible to get submodel in the selected position for the selected asset type and the the selected assetId");
             }
             return subModel;
@@ -387,69 +207,30 @@ public class AasService extends BaseService {
                     "It was not possible to search submodel!");
         }
     }
-
     /**
-     * Searches a {@code Submodel} from a Digital Twin from a given position index containing the specified parameters.
+     * Searches a {@code Submodel} from a Digital Twin by the configured semanticIds
      * <p>
      * @param   digitalTwin
      *          the {@code DigitalTwin} object with its submodels.
-     * @param   position
-     *          the {@code Integer} position index of the intended submodel.
-     * @param   registryUrl
-     *          the {@code String} registry URL of the Digital Twin.
-     * @param   edr
-     *          the {@code DataPlaneEndpoint} object with the data plane needed parameters.
      *
-     * @return a {@code Submodel} object from the position index, if exists.
+     * @return a {@code Submodel} object with the submodel found, if exists.
      *
      * @throws  ServiceException
-     *           if unable to find a {@code Submodel} for the specified position index.
+     *           if unable to find a the {@code Submodel} for the given semantic id.
      */
-    public SubModel searchSubModel(DigitalTwin digitalTwin, Integer position, String registryUrl, DataPlaneEndpoint edr) {
+    public SubModel searchSubModelBySemanticId(DigitalTwin digitalTwin) {
         try {
-            SubModel subModel = this.getSubModel(digitalTwin, position, registryUrl, edr);
+            SubModel subModel = this.getSubModelBySemanticId(digitalTwin);
             if (subModel == null) {
-                throw new ServiceException(this.getClass().getName() + "." + "searchSubModel",
+                throw new ServiceException(this.getClass().getName() + "." + "searchSubModelBySemanticId",
                         "It was not possible to get submodel in the selected position for the selected asset type and the the selected assetId");
             }
             return subModel;
         } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "searchSubModel",
+            throw new ServiceException(this.getClass().getName() + "." + "searchSubModelById",
                     e,
                     "It was not possible to search submodel!");
         }
-    }
-
-    /**
-     * Gets the {@code DigitalTwin3} of a given id.
-     * <p>
-     * @param   digitalTwinId
-     *          the {@code String} identification of the digital twin.
-     * @param   registryUrl
-     *          the {@code String} registry URL of the Digital Twin.
-     * @param   edr
-     *          the {@code DataPlaneEndpoint} object with the data plane needed parameters.
-     *
-     * @return a {@code DigitalTwin3} object with the given id, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a {@code DigitalTwin3} for the specified id.
-     */
-    public DigitalTwin3 getDigitalTwin3(String digitalTwinId, String registryUrl, DataPlaneEndpoint edr) {
-        try {
-            String path = this.getPathEndpoint(registryUrl, "digitalTwin");
-            String url = this.getRegistryUrl(registryUrl) + path + "/" + CrypUtil.toBase64Url(digitalTwinId);
-            Map<String, Object> params = httpUtil.getParams();
-            HttpHeaders headers = this.getTokenHeader(edr);
-            ResponseEntity<?> response = httpUtil.doGet(url, String.class, headers, params, true, false);
-            String responseBody = (String) response.getBody();
-            return (DigitalTwin3) jsonUtil.bindJsonNode(jsonUtil.toJsonNode(responseBody), DigitalTwin3.class);
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getDigitalTwin",
-                    e,
-                    "It was not possible to get digital twin!");
-        }
-
     }
 
     /**
@@ -470,7 +251,7 @@ public class AasService extends BaseService {
     public DigitalTwin getDigitalTwin(String digitalTwinId, String registryUrl, DataPlaneEndpoint edr) {
         try {
             String path = this.getPathEndpoint(registryUrl, "digitalTwin");
-            String url = this.getRegistryUrl(registryUrl) + path + "/" + digitalTwinId;
+            String url = this.getRegistryUrl(registryUrl) + path + "/" + CrypUtil.toBase64Url(digitalTwinId);
             Map<String, Object> params = httpUtil.getParams();
             HttpHeaders headers = this.getTokenHeader(edr);
             ResponseEntity<?> response = httpUtil.doGet(url, String.class, headers, params, true, false);
@@ -484,127 +265,26 @@ public class AasService extends BaseService {
 
     }
 
-
     /**
-     * Gets the {@code Submodel} from a Digital Twin from a given position index.
+     * Gets the {@code Submodel} from a Digital Twin by the configured semantic aspects
      * <p>
      * @param   digitalTwin
      *          the {@code DigitalTwin} object with its submodels.
-     * @param   position
-     *          the {@code Integer} position index of the intended submodel.
      *
-     * @return a {@code Submodel} object from the position index, if exists.
+     * @return a {@code Submodel} object with the submodel found, if exists.
      *
      * @throws  ServiceException
-     *           if unable to find a {@code Submodel} for the specified position index.
+     *           if unable to find a the {@code Submodel} for the given semantic id.
      */
-    public SubModel getSubModelFromDigitalTwin(DigitalTwin digitalTwin, Integer position) {
+
+    public SubModel getSubModelBySemanticId(DigitalTwin digitalTwin) {
         try {
             ArrayList<SubModel> subModels = digitalTwin.getSubmodelDescriptors();
-            if (position > subModels.size()) {
-                throw new ServiceException(this.getClass().getName() + "." + "getSubModelFromDigitalTwin",
-                        "Position selected for subModel is out of range!");
-            }
-            return subModels.get(position);
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModelFromDigitalTwin",
-                    e,
-                    "It was not possible to get subModel from digital twin!");
-        }
-    }
-
-    /**
-     * Gets the {@code Submodel} from a Digital Twin from a given position index containing the specified parameters.
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin} object with its submodels.
-     * @param   position
-     *          the {@code Integer} position index of the intended submodel.
-     * @param   registryUrl
-     *          the {@code String} registry URL of the Digital Twin.
-     * @param   edr
-     *          the {@code DataPlaneEndpoint} object with the data plane needed parameters.
-     *
-     * @return a {@code Submodel} object from the position index, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a {@code Submodel} for the specified position index.
-     */
-    public SubModel getSubModel(DigitalTwin digitalTwin, Integer position, String registryUrl, DataPlaneEndpoint edr) {
-        try {
-            String path = this.getPathEndpoint(registryUrl, "digitalTwin");
-            SubModel subModel = this.getSubModelFromDigitalTwin(digitalTwin, position);
-            String url = this.getRegistryUrl(registryUrl) + path + "/" + digitalTwin.getIdentification() + this.getPathEndpoint(registryUrl,"subModel") + "/" + subModel.getIdentification();
-            Map<String, Object> params = httpUtil.getParams();
-            HttpHeaders headers = this.getTokenHeader(edr);
-            LogUtil.printMessage(jsonUtil.toJson(headers, true));
-            ResponseEntity<?> response = httpUtil.doGet(url, String.class, headers, params, true, false);
-            String responseBody = (String) response.getBody();
-            return (SubModel) jsonUtil.bindJsonNode(jsonUtil.toJsonNode(responseBody), SubModel.class);
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModel",
-                    e,
-                    "It was not possible to get subModel!");
-        }
-    }
-
-    /**
-     * Gets the {@code Submodel3} from a Digital Twin by a given short identification.
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin3} object with its submodels.
-     * @param   idShort
-     *          the {@code String} short id of the intended submodel.
-     *
-     * @return a {@code Submodel3} object with the submodel found, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a the {@code Submodel3} for the given id.
-     */
-    public SubModel3 getSubModel3ById(DigitalTwin3 digitalTwin, String idShort) {
-        try {
-            ArrayList<SubModel3> subModels = digitalTwin.getSubmodelDescriptors();
             if (subModels.size() < 1) {
-                throw new ServiceException(this.getClass().getName() + "." + "getSubModel3ById",
+                throw new ServiceException(this.getClass().getName() + "." + "getSubModelBySemanticId",
                         "No subModel found in digitalTwin!");
             }
-            // Search for first subModel with matching idShort, if it fails gives null
-            SubModel3 subModel = subModels.stream().filter(s -> s.getIdShort().equalsIgnoreCase(idShort)).findFirst().orElse(null);
-
-            if (subModel == null) {
-                // If the subModel idShort does not exist
-                throw new ServiceException(this.getClass().getName() + "." + "getSubModel3ById",
-                        "SubModel for idShort not found!");
-            }
-            // Return subModel if found
-            return subModel;
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModel3ById",
-                    e,
-                    "It was not possible to get subModel!");
-        }
-    }
-
-    /**
-     * Gets the {@code Submodel3} from a Digital Twin by the configured semantic aspects
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin3} object with its submodels.
-     *
-     * @return a {@code Submodel3} object with the submodel found, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a the {@code Submodel3} for the given semantic id.
-     */
-
-    public SubModel3 getSubModel3BySemanticId(DigitalTwin3 digitalTwin) {
-        try {
-            ArrayList<SubModel3> subModels = digitalTwin.getSubmodelDescriptors();
-            if (subModels.size() < 1) {
-                throw new ServiceException(this.getClass().getName() + "." + "getSubModel3BySemanticId",
-                        "No subModel found in digitalTwin!");
-            }
-            SubModel3 subModel = null;
+            SubModel subModel = null;
             // Search for first subModel with matching semanticId, if it fails gives null
             for (String semanticId: passportConfig.getAspects()) {
                 subModel = subModels.stream().filter(s -> s.getSemanticId().getKeys().stream().filter(k -> k.getType().equalsIgnoreCase(submodelTypeKey) && k.getValue().equalsIgnoreCase(semanticId)) != null).findFirst().orElse(null);
@@ -613,82 +293,45 @@ public class AasService extends BaseService {
                 }
             }
             // If the subModel semanticId does not exist
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModel3BySemanticId",
+            throw new ServiceException(this.getClass().getName() + "." + "getSubModelBySemanticId",
                     "SubModel for SemanticId not found!");
         } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModel3BySemanticId",
+            throw new ServiceException(this.getClass().getName() + "." + "getSubModelBySemanticId",
                     e,
                     "It was not possible to get subModel!");
         }
     }
     /**
-     * Gets the {@code Submodel3} from a Digital Twin by a given semantic identification.
+     * Gets the {@code Submodel} from a Digital Twin by a given semantic identification.
      * <p>
      * @param   digitalTwin
-     *          the {@code DigitalTwin3} object with its submodels.
+     *          the {@code DigitalTwin} object with its submodels.
      * @param   semanticId
      *          the {@code String} semantic id of the intended submodel.
      *
-     * @return a {@code Submodel3} object with the submodel found, if exists.
+     * @return a {@code Submodel} object with the submodel found, if exists.
      *
      * @throws  ServiceException
-     *           if unable to find a the {@code Submodel3} for the given semantic id.
+     *           if unable to find a the {@code Submodel} for the given semantic id.
      */
-    public SubModel3 getSubModel3BySemanticId(DigitalTwin3 digitalTwin, String semanticId) {
+    public SubModel getSubModelBySemanticId(DigitalTwin digitalTwin, String semanticId) {
         try {
-            ArrayList<SubModel3> subModels = digitalTwin.getSubmodelDescriptors();
+            ArrayList<SubModel> subModels = digitalTwin.getSubmodelDescriptors();
             if (subModels.size() < 1) {
-                throw new ServiceException(this.getClass().getName() + "." + "getSubModel3BySemanticId",
+                throw new ServiceException(this.getClass().getName() + "." + "getSubModelBySemanticId",
                         "No subModel found in digitalTwin!");
             }
-            SubModel3 subModel = null;
+            SubModel subModel = null;
             // Search for first subModel with matching semanticId, if it fails gives null
             subModel = subModels.stream().filter(s -> s.getSemanticId().getKeys().stream().filter(k -> k.getType().equalsIgnoreCase(submodelTypeKey) && k.getValue().equalsIgnoreCase(semanticId)) != null).findFirst().orElse(null);
             if (subModel != null) {
                 return subModel; // Return subModel if found
             }
             // If the subModel semanticId does not exist
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModel3BySemanticId",
+            throw new ServiceException(this.getClass().getName() + "." + "getSubModelBySemanticId",
                     "SubModel for SemanticId not found!");
         } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModel3BySemanticId",
-                    e,
-                    "It was not possible to get subModel!");
-        }
-    }
-
-    /**
-     * Gets the {@code Submodel} from a Digital Twin by a given short identification.
-     * <p>
-     * @param   digitalTwin
-     *          the {@code DigitalTwin} object with its submodels.
-     * @param   idShort
-     *          the {@code String} short id of the intended submodel.
-     *
-     * @return a {@code Submodel} object with the submodel found, if exists.
-     *
-     * @throws  ServiceException
-     *           if unable to find a the {@code Submodel} for the given id.
-     */
-    public SubModel getSubModelById(DigitalTwin digitalTwin, String idShort) {
-        try {
-            ArrayList<SubModel> subModels = digitalTwin.getSubmodelDescriptors();
-            if (subModels.size() < 1) {
-                throw new ServiceException(this.getClass().getName() + "." + "getSubModelByIdShort",
-                        "No subModel found in digitalTwin!");
-            }
-            // Search for first subModel with matching idShort, if it fails gives null
-            SubModel subModel = subModels.stream().filter(s -> s.getIdShort().equalsIgnoreCase(idShort)).findFirst().orElse(null);
-
-            if (subModel == null) {
-                // If the subModel idShort does not exist
-                throw new ServiceException(this.getClass().getName() + "." + "getSubModelByIdShort",
-                        "SubModel for idShort not found!");
-            }
-            // Return subModel if found
-            return subModel;
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "getSubModelByIdShort",
+            throw new ServiceException(this.getClass().getName() + "." + "getSubModelBySemanticId",
                     e,
                     "It was not possible to get subModel!");
         }
@@ -707,8 +350,7 @@ public class AasService extends BaseService {
      */
     public HttpHeaders getTokenHeader(DataPlaneEndpoint edr) {
         try {
-            // If the dtr is central we just need the token
-            if (this.central || edr == null) {
+            if (edr == null) {
                 // In case it fails we should throw get the token
                 JwtToken token = authService.getToken();
                 return this.httpUtil.getHeadersWithToken(token.getAccessToken());
@@ -734,69 +376,7 @@ public class AasService extends BaseService {
      *
      */
     public String getRegistryUrl(String url) {
-            return (url == null || this.central || url.isEmpty()) ? this.registryUrl : url;
-    }
-
-    /**
-     * Searches for centralized Digital Twin Registries.
-     * <p>
-     * @param   processId
-     *          the {@code String} identification of the process.
-     * @param   searchBody
-     *          the {@code Search} object with the Request Body of the HTTP request.
-     *
-     * @return an {@code AssetSearch} object with the connector address and the asset id of the DTR.
-     *
-     * @throws  ServiceException
-     *           if unable to search and/or find the DTR.
-     */
-    public AssetSearch centralDtrSearch(String processId, Search searchBody){
-        try {
-            // Start Digital Twin Query
-            AasService.DigitalTwinRegistryQueryById digitalTwinRegistry = this.new DigitalTwinRegistryQueryById(searchBody);
-            Long dtRequestTime = DateTimeUtil.getTimestamp();
-            Thread digitalTwinRegistryThread = ThreadUtil.runThread(digitalTwinRegistry);
-
-            // Wait for digital twin query
-            digitalTwinRegistryThread.join();
-            DigitalTwin digitalTwin;
-            SubModel subModel;
-            String connectorId;
-            String connectorAddress;
-            try {
-                digitalTwin = digitalTwinRegistry.getDigitalTwin();
-                subModel = digitalTwinRegistry.getSubModel();
-                connectorId = subModel.getIdShort();
-                EndPoint endpoint = subModel.getEndpoints().stream().filter(obj -> obj.getInterfaceName().equals(dtrConfig.getEndpointInterface())).findFirst().orElse(null);
-                if (endpoint == null) {
-                    throw new ControllerException(this.getClass().getName(), "No EDC endpoint found in DTR SubModel!");
-                }
-                connectorAddress = endpoint.getProtocolInformation().getEndpointAddress();
-            } catch (Exception e) {
-                return null;
-            }
-            if (connectorId.isEmpty() || connectorAddress.isEmpty()) {
-                return null;
-            }
-
-
-            try {
-                connectorAddress = CatenaXUtil.buildEndpoint(connectorAddress);
-            } catch (Exception e) {
-                return null;
-            }
-            if (connectorAddress.isEmpty()) {
-                return null;
-            }
-            processManager.saveDigitalTwin(processId, digitalTwin, dtRequestTime);
-            LogUtil.printDebug("[PROCESS " + processId + "] Digital Twin [" + digitalTwin.getIdentification() + "] and Submodel [" + subModel.getIdentification() + "] with EDC endpoint [" + connectorAddress + "] retrieved from DTR");
-            String assetId = String.join("-", digitalTwin.getIdentification(), subModel.getIdentification());
-            return new AssetSearch(assetId, connectorAddress);
-        } catch (Exception e) {
-            throw new ServiceException(this.getClass().getName() + "." + "centralDtrSearch",
-                    e,
-                    "It was not possible to search and find digital twin");
-        }
+            return url;
     }
 
     /**
@@ -873,7 +453,7 @@ public class AasService extends BaseService {
             String url = this.getRegistryUrl(registryUrl) + path;
             Map<String, Object> params = httpUtil.getParams();
             ResponseEntity<?> response = null;
-            if (!this.central && registryUrl != null && edr != null) {
+            if (registryUrl != null && edr != null) {
                 // Set request body as post if the central query is disabled
                 // Query as GET if the central query is enabled
                 Map<String, ?> assetIds = Map.of(
@@ -965,14 +545,12 @@ public class AasService extends BaseService {
     public class DecentralDigitalTwinRegistryQueryById implements Runnable {
 
         /** ATTRIBUTES **/
-        private final String dppIdShort = "digitalProductPass";
         private DataPlaneEndpoint edr;
-        private SubModel3 subModel;
-        private DigitalTwin3 digitalTwin;
+        private SubModel subModel;
+        private DigitalTwin digitalTwin;
         private final String assetId;
         private final String idType;
         private final Integer dtIndex;
-        private final String idShort;
         private final String semanticId;
 
         /** CONSTRUCTOR(S) **/
@@ -980,27 +558,10 @@ public class AasService extends BaseService {
             this.assetId = search.getId();
             this.idType = search.getIdType();
             this.dtIndex = search.getDtIndex();
-            this.idShort = search.getIdShort();
             this.edr = edr;
             this.semanticId = search.getSemanticId();
         }
-        /**
-         * This method is exclusively to search for a digital twin and the submodel.
-         *
-         * <p> It's a Thread level method from Runnable interface and does the {@code DigitalTwin3} and {@code Submodel3} search, setting the results
-         * to this class object.
-         * <p> The submodel search it's done by the idShort or semanticId parameters depending on if the semantic Id is available or not, respectively.
-         *
-         */
-        @Override
-        public void run() {
-            this.setDigitalTwin(searchDigitalTwin3(this.getIdType(), this.getAssetId(), this.getDtIndex(),  this.getEdr().getEndpoint(), this.getEdr()));
-            if(semanticId.isEmpty()){
-                this.setSubModel(searchSubModel3BySemanticId(this.getDigitalTwin()));
-            }else {
-                this.setSubModel(searchSubModel3BySemanticId(this.getDigitalTwin(), semanticId));
-            }
-        }
+
         /** GETTERS AND SETTERS **/
         public DataPlaneEndpoint getEdr() {
             return edr;
@@ -1008,63 +569,14 @@ public class AasService extends BaseService {
         public void setEdr(DataPlaneEndpoint edr) {
             this.edr = edr;
         }
-        public SubModel3 getSubModel() {
-            return subModel;
-        }
-        public void setSubModel(SubModel3 subModel) {
-            this.subModel = subModel;
-        }
-        public DigitalTwin3 getDigitalTwin() {
-            return digitalTwin;
-        }
-        public void setDigitalTwin(DigitalTwin3 digitalTwin) {
-            this.digitalTwin = digitalTwin;
-        }
-        public String getAssetId() {
-            return assetId;
-        }
-        public String getIdType() {
-            return idType;
-        }
-        public Integer getDtIndex() {
-            return dtIndex;
-        }
-        public String getIdShort() {
-            return idShort;
-        }
-        public String getSemanticId() { return semanticId; }
-
-        /** METHODS **/
-
-    }
-
-    public class DigitalTwinRegistryQueryById implements Runnable {
-
-        /** ATTRIBUTES **/
-        private SubModel subModel;
-        private DigitalTwin digitalTwin;
-        private final String assetId;
-        private final String idType;
-        private final Integer dtIndex;
-        private final String idShort;
-
-        /** CONSTRUCTOR(S) **/
-        public DigitalTwinRegistryQueryById(Search search) {
-            this.assetId = search.getId();
-            this.idType = search.getIdType();
-            this.dtIndex = search.getDtIndex();
-            this.idShort = search.getIdShort();
-        }
-
-        /** GETTERS AND SETTERS **/
         public SubModel getSubModel() {
-            return this.subModel;
-        }
-        public DigitalTwin getDigitalTwin() {
-            return this.digitalTwin;
+            return subModel;
         }
         public void setSubModel(SubModel subModel) {
             this.subModel = subModel;
+        }
+        public DigitalTwin getDigitalTwin() {
+            return digitalTwin;
         }
         public void setDigitalTwin(DigitalTwin digitalTwin) {
             this.digitalTwin = digitalTwin;
@@ -1078,72 +590,25 @@ public class AasService extends BaseService {
         public Integer getDtIndex() {
             return dtIndex;
         }
-        public String getIdShort() {
-            return idShort;
-        }
+        public String getSemanticId() { return semanticId; }
 
         /** METHODS **/
-
         /**
          * This method is exclusively to search for a digital twin and the submodel.
          *
          * <p> It's a Thread level method from Runnable interface and does the {@code DigitalTwin} and {@code Submodel} search, setting the results
          * to this class object.
-         * <p> The submodel search it's done by the idShort.
+         * <p> The submodel search it's done by semanticId parameter depending on if the semanticId is available or not, if not does the search by a semanticId default value.
          *
          */
         @Override
         public void run() {
-            this.digitalTwin = searchDigitalTwin(this.idType, this.assetId, this.dtIndex, null, null);
-            this.subModel = searchSubModelById(this.digitalTwin, this.idShort);
+            this.setDigitalTwin(searchDigitalTwin(this.getIdType(), this.getAssetId(), this.getDtIndex(),  this.getEdr().getEndpoint(), this.getEdr()));
+            if(semanticId.isEmpty()){
+                this.setSubModel(searchSubModelBySemanticId(this.getDigitalTwin()));
+            }else {
+                this.setSubModel(searchSubModelBySemanticId(this.getDigitalTwin(), semanticId));
+            }
         }
-
     }
-
-    @SuppressWarnings("Unused")
-    public class DigitalTwinRegistryQuery implements Runnable {
-
-        /** ATTRIBUTES **/
-        private SubModel subModel;
-        private DigitalTwin digitalTwin;
-        private final String assetId;
-        private final String idType;
-        private final Integer dtIndex;
-
-        /** CONSTRUCTOR(S) **/
-        @SuppressWarnings("Unused")
-        public DigitalTwinRegistryQuery(String assetId, String idType, Integer dtIndex) {
-            this.assetId = assetId;
-            this.idType = idType;
-            this.dtIndex = dtIndex;
-        }
-
-        /** GETTERS AND SETTERS **/
-        @SuppressWarnings("Unused")
-        public SubModel getSubModel() {
-            return this.subModel;
-        }
-        @SuppressWarnings("Unused")
-        public DigitalTwin getDigitalTwin() {
-            return this.digitalTwin;
-        }
-
-        /** METHODS **/
-
-        /**
-         * This method is exclusively to search for a digital twin and the submodel.
-         *
-         * <p> It's a Thread level method from Runnable interface and does the {@code DigitalTwin} and {@code Submodel} search, setting the results
-         * to this class object.
-         * <p> The submodel search it's done by index position.
-         *
-         */
-        @Override
-        public void run() {
-            this.digitalTwin = searchDigitalTwin(this.idType, this.assetId, this.dtIndex, null, null);
-            this.subModel = searchSubModel(this.digitalTwin, this.dtIndex, null, null);
-        }
-
-    }
-
 }
